Fix: 분류기 프롬프트 수정 및 코드 보수(#19)
